requirements.txt
# Core API
fastapi==0.115.12
uvicorn[standard]==0.34.2
gunicorn==23.0.0
python-multipart==0.0.20
pydantic-settings==2.9.1

# ML / Inference
open-clip-torch==2.30.0
torch==2.6.0
torchvision==0.21.0
Pillow==11.2.1

# Caching & Queue
redis==5.3.0
celery[redis]==5.5.1

# Utilities
python-dotenv==1.1.0

# Testing
pytest==8.3.5
httpx==0.28.1
pytest-asyncio==0.26.0
